% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_optimization.R
\name{add_performance_metrics}
\alias{add_performance_metrics}
\title{Add performance metrics to optimal locations}
\usage{
add_performance_metrics(optimal_locations, accidents)
}
\description{
Add performance metrics to optimal locations
}
\keyword{internal}
