% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero_simple_stars.R
\name{apply_focal_mean}
\alias{apply_focal_mean}
\title{Apply Focal Mean}
\usage{
apply_focal_mean(data, window_size)
}
\arguments{
\item{data}{matrix, input data}

\item{window_size}{numeric, window size}
}
\value{
matrix, smoothed data
}
\description{
Applies focal mean to a matrix.
}
\keyword{internal}
