% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero_stars_enhanced.R
\name{apply_kernel_parallel}
\alias{apply_kernel_parallel}
\title{Apply Kernel Convolution (Parallel)}
\usage{
apply_kernel_parallel(raster_data, kernel)
}
\arguments{
\item{raster_data}{stars object with raster data}

\item{kernel}{matrix, convolution kernel}
}
\value{
stars object with convolved data
}
\description{
Applies kernel convolution to raster data using parallel processing.
}
\keyword{internal}
