% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero_stars.R
\name{apply_kernel_sequential}
\alias{apply_kernel_sequential}
\title{Apply Kernel Sequential}
\usage{
apply_kernel_sequential(template, accident_coords, bandwidth)
}
\arguments{
\item{template}{stars template}

\item{accident_coords}{accident coordinates}

\item{bandwidth}{kernel bandwidth}
}
\value{
stars object with density
}
\description{
Apply Kernel Sequential
}
