% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing_analysis.R
\name{calculate_euclidean_route}
\alias{calculate_euclidean_route}
\title{Calculate Euclidean distance route (straight line)}
\usage{
calculate_euclidean_route(new_accident, optimal_locations)
}
\description{
Calculate Euclidean distance route (straight line)
}
\keyword{internal}
