% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing_analysis.R
\name{calculate_hybrid_route}
\alias{calculate_hybrid_route}
\title{Calculate hybrid route (weighted combination)}
\usage{
calculate_hybrid_route(new_accident, optimal_locations, data)
}
\description{
Calculate hybrid route (weighted combination)
}
\keyword{internal}
