% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing_analysis.R
\name{calculate_road_network_route}
\alias{calculate_road_network_route}
\title{Calculate road network route (simplified)}
\usage{
calculate_road_network_route(new_accident, optimal_locations, data)
}
\description{
Calculate road network route (simplified)
}
\keyword{internal}
