% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing_analysis.R
\name{calculate_route_statistics}
\alias{calculate_route_statistics}
\title{Calculate route statistics}
\usage{
calculate_route_statistics(route_result)
}
\description{
Calculate route statistics
}
\keyword{internal}
