% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero_stars_enhanced.R
\name{convolve_2d}
\alias{convolve_2d}
\title{2D Convolution Function}
\usage{
convolve_2d(data, kernel)
}
\arguments{
\item{data}{matrix, input data}

\item{kernel}{matrix, convolution kernel}
}
\value{
matrix, convolved data
}
\description{
Performs 2D convolution operation on matrix data.
}
\keyword{internal}
