% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero_optimal.R
\name{create_empty_optimal_locations}
\alias{create_empty_optimal_locations}
\title{Create empty optimal locations object}
\usage{
create_empty_optimal_locations()
}
\value{
empty sero_optimal_locations object
}
\description{
Create empty optimal locations object
}
