% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing_analysis.R
\name{create_euclidean_route}
\alias{create_euclidean_route}
\title{Create euclidean route as fallback}
\usage{
create_euclidean_route(accident_location, nearest_base)
}
\description{
Create euclidean route as fallback
}
\keyword{internal}
