% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing_analysis.R
\name{create_interactive_routing_map}
\alias{create_interactive_routing_map}
\title{Create interactive map for emergency routing}
\usage{
create_interactive_routing_map(
  data,
  optimal_locations,
  districts,
  include_roads = TRUE
)
}
\arguments{
\item{data}{Original data list}

\item{optimal_locations}{Optimal locations in WGS84}

\item{districts}{Districts in WGS84}

\item{include_roads}{Whether to show roads}
}
\value{
Interactive leaflet map
}
\description{
Create interactive map for emergency routing
}
