% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_analysis.R
\name{create_optimal_visualization}
\alias{create_optimal_visualization}
\title{Create visualization for optimal locations}
\usage{
create_optimal_visualization(
  accidents,
  districts,
  optimal_locations,
  data,
  method
)
}
\description{
Create visualization for optimal locations
}
\keyword{internal}
