% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing_analysis.R
\name{create_route_visualization}
\alias{create_route_visualization}
\title{Create route visualization}
\usage{
create_route_visualization(
  new_accident,
  optimal_locations,
  districts,
  route_result,
  data,
  include_roads
)
}
\description{
Create route visualization
}
\keyword{internal}
