% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/professional_mapping.R
\name{create_static_accident_plot}
\alias{create_static_accident_plot}
\title{Create static ggplot for accidents}
\usage{
create_static_accident_plot(accidents, districts, landuse, severity_column)
}
\description{
Create static ggplot for accidents
}
\keyword{internal}
