% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/professional_mapping.R
\name{create_static_optimal_plot}
\alias{create_static_optimal_plot}
\title{Create static ggplot for optimal locations}
\usage{
create_static_optimal_plot(optimal_locs, districts, roads, landuse, accidents)
}
\description{
Create static ggplot for optimal locations
}
\keyword{internal}
