% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero_stars.R
\name{extract_max_density}
\alias{extract_max_density}
\title{Extract Maximum Density}
\usage{
extract_max_density(polygons, density_surface)
}
\arguments{
\item{polygons}{sf object with hotspot polygons}

\item{density_surface}{stars object with density surface}
}
\value{
numeric vector of maximum density values
}
\description{
Extract Maximum Density
}
