% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_analysis.R
\name{find_density_locations}
\alias{find_density_locations}
\title{Density-based optimal location finding}
\usage{
find_density_locations(coords, districts, k)
}
\description{
Density-based optimal location finding
}
\keyword{internal}
