% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_analysis.R
\name{find_grid_locations}
\alias{find_grid_locations}
\title{Grid-based optimal location finding}
\usage{
find_grid_locations(coords, districts, k)
}
\description{
Grid-based optimal location finding
}
\keyword{internal}
