% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_analysis.R
\name{find_hybrid_locations}
\alias{find_hybrid_locations}
\title{Hybrid optimization approach}
\usage{
find_hybrid_locations(coords, districts, k)
}
\description{
Hybrid optimization approach
}
\keyword{internal}
