% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_analysis.R
\name{find_kmeans_locations}
\alias{find_kmeans_locations}
\title{K-means clustering for optimal locations}
\usage{
find_kmeans_locations(coords, k)
}
\description{
K-means clustering for optimal locations
}
\keyword{internal}
