% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero_styling.R
\name{get_sero_scale}
\alias{get_sero_scale}
\title{Get SERO Color Scale}
\usage{
get_sero_scale(type = "density", alpha = 0.7)
}
\arguments{
\item{type}{character, type of scale ("density", "accidents", "hotspots")}

\item{alpha}{numeric, transparency level (0-1)}
}
\description{
Returns standardized color scale for density visualizations
}
