% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero_stars_enhanced.R
\name{plot.sero_hotspots_stars}
\alias{plot.sero_hotspots_stars}
\title{Plot Method for Stars-based Hotspots}
\usage{
\method{plot}{sero_hotspots_stars}(x, ...)
}
\arguments{
\item{x}{sero_hotspots_stars object}

\item{...}{additional arguments passed to ggplot}
}
\value{
ggplot2 object
}
\description{
Plots stars-based hotspots with density surface.
}
