% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero_enhanced_stars.R
\name{plot.sero_optimal_locations_enhanced_stars}
\alias{plot.sero_optimal_locations_enhanced_stars}
\title{Standardized Plot Method for Enhanced Stars Optimization}
\usage{
\method{plot}{sero_optimal_locations_enhanced_stars}(x, ...)
}
\arguments{
\item{x}{sero_optimal_locations_enhanced_stars object}

\item{...}{additional arguments}
}
\value{
ggplot2 object
}
\description{
Standardized Plot Method for Enhanced Stars Optimization
}
