% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero_simple_stars.R
\name{plot.sero_optimal_locations_simple_stars}
\alias{plot.sero_optimal_locations_simple_stars}
\title{Plot Method for Simple Stars Optimization}
\usage{
\method{plot}{sero_optimal_locations_simple_stars}(x, ...)
}
\arguments{
\item{x}{sero_optimal_locations_simple_stars object}

\item{...}{additional arguments}
}
\value{
ggplot2 object
}
\description{
Plot Method for Simple Stars Optimization
}
