% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero_stars_optimization.R
\name{plot.sero_optimal_locations_stars}
\alias{plot.sero_optimal_locations_stars}
\title{Plot Method for Stars-based Optimal Locations}
\usage{
\method{plot}{sero_optimal_locations_stars}(x, ...)
}
\arguments{
\item{x}{sero_optimal_locations_stars object}

\item{...}{additional arguments passed to ggplot}
}
\value{
ggplot2 object
}
\description{
Plots optimal locations with density surface and performance metrics.
}
