% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero_interactive.R
\name{plot.sero_route_to_accident}
\alias{plot.sero_route_to_accident}
\title{Plot method for sero_route_to_accident}
\usage{
\method{plot}{sero_route_to_accident}(x, ...)
}
\arguments{
\item{x}{sero_route_to_accident object}

\item{...}{additional arguments (unused)}
}
\value{
ggplot2 object
}
\description{
Plot method for sero_route_to_accident
}
