% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero_optimal.R
\name{print.sero_optimal_locations}
\alias{print.sero_optimal_locations}
\title{Print method for sero_optimal_locations}
\usage{
\method{print}{sero_optimal_locations}(x, ...)
}
\arguments{
\item{...}{additional arguments (unused)}

\item{object}{sero_optimal_locations object}
}
\description{
Print method for sero_optimal_locations
}
