% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero_interactive.R
\name{print.sero_route_to_accident}
\alias{print.sero_route_to_accident}
\title{Print method for sero_route_to_accident}
\usage{
\method{print}{sero_route_to_accident}(x, ...)
}
\arguments{
\item{x}{sero_route_to_accident object}

\item{...}{additional arguments (unused)}
}
\description{
Print method for sero_route_to_accident
}
