% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero_routes.R
\name{print.sero_routes}
\alias{print.sero_routes}
\title{Print method for sero_routes}
\usage{
\method{print}{sero_routes}(x, ...)
}
\arguments{
\item{x}{sero_routes object}

\item{...}{additional arguments (unused)}
}
\description{
Print method for sero_routes
}
