% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/professional_mapping.R
\name{select_optimal_locations}
\alias{select_optimal_locations}
\title{Select optimal locations from scored grid}
\usage{
select_optimal_locations(scores, num_locations)
}
\description{
Select optimal locations from scored grid
}
\keyword{internal}
