% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero_routes.R
\name{sero_calculate_routes}
\alias{sero_calculate_routes}
\title{Calculate fastest routes to accident scenes}
\usage{
sero_calculate_routes(locations, accidents, max_routes = 10)
}
\arguments{
\item{locations}{sero_optimal_locations object or sf object with service locations}

\item{accidents}{sf object with accident locations}

\item{max_routes}{Maximum number of routes to calculate (default=10)}
}
\value{
sero_routes S3 object
}
\description{
Computes fastest routes from emergency service locations to accident scenes.
Uses simple straight-line distance calculations for speed.
}
\examples{
\dontrun{
data <- sero_load_data()
locations <- sero_compute_optimal_locations(data)
routes <- sero_calculate_routes(locations, data$accident)
plot(routes)
}
}
