% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero_styling.R
\docType{data}
\name{SERO_COLORS}
\alias{SERO_COLORS}
\title{SERO Color Palette}
\format{
An object of class \code{list} of length 15.
}
\usage{
SERO_COLORS
}
\description{
Standardized color scheme for SERO package visualizations
}
\keyword{datasets}
