% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero_interactive.R
\name{sero_create_interactive_map}
\alias{sero_create_interactive_map}
\title{Interactive accident selection and routing for SERO package}
\usage{
sero_create_interactive_map(data, optimal_locations, max_accidents = 100)
}
\arguments{
\item{data}{List containing spatial data layers}

\item{optimal_locations}{sero_optimal_locations object with calculated optimal locations}

\item{max_accidents}{Maximum number of accidents to display (default=100)}
}
\value{
Interactive plotly map
}
\description{
This file contains functions for interactive accident selection and routing
from optimal emergency service locations to selected accident sites.
Create an interactive map for accident selection
}
\details{
Creates an interactive map where users can click on accident locations
to select them for routing analysis. This function uses plotly for interactivity.
}
\examples{
\dontrun{
data <- sero_load_data()
optimal_locs <- sero_calculate_optimal_locations(data)
interactive_map <- sero_create_interactive_map(data, optimal_locs)
interactive_map
}
}
