% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspot_analysis.R
\name{sero_interactive_hotspots}
\alias{sero_interactive_hotspots}
\title{Create interactive hotspot map}
\usage{
sero_interactive_hotspots(hotspot_result, data)
}
\arguments{
\item{hotspot_result}{Result from sero_hotspot_analysis()}

\item{data}{Original data list}
}
\value{
leaflet map object
}
\description{
Create interactive hotspot map
}
