% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero_optimal.R
\name{sero_plot_all}
\alias{sero_plot_all}
\title{Create comprehensive map with all layers}
\usage{
sero_plot_all(
  optimal_locations,
  data,
  show_grid = FALSE,
  show_munster = TRUE,
  show_hotspots = TRUE,
  show_roads = TRUE,
  show_landuse = FALSE,
  show_population = FALSE
)
}
\arguments{
\item{optimal_locations}{sero_optimal_locations object}

\item{data}{Original data containing all spatial layers}

\item{show_grid}{Logical, whether to show grid cells (default=FALSE)}

\item{show_munster}{Logical, whether to show Muenster city boundaries (default=TRUE)}

\item{show_hotspots}{Logical, whether to show accident hotspots (default=TRUE)}

\item{show_roads}{Logical, whether to show road network (default=TRUE)}

\item{show_landuse}{Logical, whether to show land use (default=FALSE)}

\item{show_population}{Logical, whether to show population density (default=FALSE)}
}
\value{
ggplot2 object
}
\description{
Create comprehensive map with all layers
}
