% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero_visualization.R
\name{sero_plot_combined}
\alias{sero_plot_combined}
\title{Combine multiple plots into a single visualization}
\usage{
sero_plot_combined(
  data,
  include_layers = c("districts", "landuse", "roads", "accidents"),
  alpha_background = 0.3
)
}
\arguments{
\item{data}{List containing spatial data layers}

\item{include_layers}{Vector of layer names to include}

\item{alpha_background}{Alpha value for background layers (default=0.3)}
}
\value{
ggplot2 object
}
\description{
Combine multiple plots into a single visualization
}
