% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero_visualization.R
\name{sero_plot_landuse}
\alias{sero_plot_landuse}
\title{Plot land use data with labels}
\usage{
sero_plot_landuse(data, label_landuse = TRUE)
}
\arguments{
\item{data}{List containing spatial data layers}

\item{label_landuse}{Logical, whether to show land use labels (default=TRUE)}
}
\value{
ggplot2 object
}
\description{
Plot land use data with labels
}
