% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero_visualization.R
\name{sero_plot_population}
\alias{sero_plot_population}
\title{Plot population density}
\usage{
sero_plot_population(data, show_values = TRUE)
}
\arguments{
\item{data}{List containing spatial data layers}

\item{show_values}{Logical, whether to show density values (default=TRUE)}
}
\value{
ggplot2 object
}
\description{
Plot population density
}
