% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero_visualization.R
\name{sero_plot_roads}
\alias{sero_plot_roads}
\title{Plot road network with customizable styling}
\usage{
sero_plot_roads(data, road_type = "all", road_border_color = "darkred")
}
\arguments{
\item{data}{List containing spatial data layers}

\item{road_type}{Character, type of roads to highlight (default="all")}

\item{road_border_color}{Color for road lines (default="darkred"). Allows users to customize the color of road borders.}
}
\value{
ggplot2 object
}
\description{
This function creates a visualization of the road network with customizable colors and road type filtering.
}
