% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero_workflow.R
\name{sero_quick_demo}
\alias{sero_quick_demo}
\title{Quick workflow for testing and demonstration}
\usage{
sero_quick_demo(interactive = TRUE)
}
\arguments{
\item{interactive}{logical whether to launch interactive routing}
}
\value{
analysis results or interactive app
}
\description{
Quick workflow for testing and demonstration
}
