% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero_interactive.R
\name{sero_route}
\alias{sero_route}
\title{Calculate route from optimal locations to selected accident}
\usage{
sero_route(optimal_locations, selected_accident, show_all_routes = FALSE)
}
\arguments{
\item{optimal_locations}{sero_optimal_locations object}

\item{selected_accident}{sf object with single accident location}

\item{show_all_routes}{Logical, whether to show routes from all optimal locations (default=FALSE)}
}
\value{
sero_route_to_accident S3 object
}
\description{
Calculates the optimal route from the nearest emergency service location
to a selected accident location.
}
