% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero_routes.R
\name{sero_save_locations}
\alias{sero_save_locations}
\title{Save optimal locations as a dataset}
\usage{
sero_save_locations(locations, filename = "optimal_locations.gpkg")
}
\arguments{
\item{locations}{sero_optimal_locations object}

\item{filename}{File path to save the locations (default: "optimal_locations.gpkg")}
}
\description{
Saves the computed optimal locations to a file for reuse in routing.
Uses GeoPackage format which is optimal for sf objects.
}
\examples{
\dontrun{
data <- sero_load_data()
locations <- sero_optimal(data)

# Save as GeoPackage (recommended - optimal for sf objects)
sero_save_locations(locations, "optimal_locations.gpkg")

# Use custom filename
sero_save_locations(locations, "my_emergency_locations.gpkg")
}
}
