% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero_optimaLoc.R
\name{sero_save_optimal_locations}
\alias{sero_save_optimal_locations}
\title{Save Optimal Locations to Database}
\usage{
sero_save_optimal_locations(locations, db_path)
}
\arguments{
\item{locations}{sf object containing optimal locations}

\item{db_path}{character, path to database file}
}
\value{
NULL (saves to file)
}
\description{
Saves optimal emergency service locations to a database or GeoPackage file.
}
\examples{
\dontrun{
optimal_locs <- sero_find_optimal_locations(accident_data)
sero_save_optimal_locations(optimal_locs, "optimal_locations.gpkg")
}
}
