% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero_interactive.R
\name{sero_select}
\alias{sero_select}
\title{Select accident by coordinates}
\usage{
sero_select(data, longitude, latitude, max_distance = 1000)
}
\arguments{
\item{data}{List containing spatial data layers}

\item{longitude}{Longitude of selected point}

\item{latitude}{Latitude of selected point}

\item{max_distance}{Maximum distance in meters to search for accidents (default=1000)}
}
\value{
Selected accident as sf object or NULL if none found
}
\description{
Selects the nearest accident to the provided coordinates. This function
is typically used after a user clicks on the interactive map.
}
