% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero_stars_optimization.R
\name{summary.sero_optimal_locations_stars}
\alias{summary.sero_optimal_locations_stars}
\title{Summary Method for Stars-based Optimal Locations}
\usage{
\method{summary}{sero_optimal_locations_stars}(object, ...)
}
\arguments{
\item{object}{sero_optimal_locations_stars object}

\item{...}{additional arguments}
}
\value{
character vector with summary
}
\description{
Provides summary statistics for optimal locations.
}
