% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero_stars.R
\name{apply_kernel_parallel}
\alias{apply_kernel_parallel}
\title{Apply Kernel Parallel}
\usage{
apply_kernel_parallel(template, accident_coords, bandwidth, cl)
}
\arguments{
\item{template}{stars template}

\item{accident_coords}{accident coordinates}

\item{bandwidth}{kernel bandwidth}

\item{cl}{cluster object}
}
\value{
stars object with density
}
\description{
Apply Kernel Parallel
}
