% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero_stars_enhanced.R
\name{apply_kernel_sequential}
\alias{apply_kernel_sequential}
\title{Apply Kernel Convolution (Sequential)}
\usage{
apply_kernel_sequential(raster_data, kernel)
}
\arguments{
\item{raster_data}{stars object with raster data}

\item{kernel}{matrix, convolution kernel}
}
\value{
stars object with convolved data
}
\description{
Applies kernel convolution to raster data using sequential processing.
}
\keyword{internal}
