% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero_routes.R
\name{calculate_straight_routes}
\alias{calculate_straight_routes}
\title{Calculate straight-line routes}
\usage{
calculate_straight_routes(service_locations, accidents, max_routes)
}
\arguments{
\item{service_locations}{sf object with service locations}

\item{accidents}{sf object with accident locations}

\item{max_routes}{Maximum number of routes to calculate}
}
\value{
List with routes sf object
}
\description{
Calculate straight-line routes
}
