% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspots.R
\name{create_empty_hotspots}
\alias{create_empty_hotspots}
\title{Create empty hotspots object}
\usage{
create_empty_hotspots()
}
\value{
empty sero_hotspots object
}
\description{
Create empty hotspots object
}
