% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero_stars.R
\name{extract_mean_density}
\alias{extract_mean_density}
\title{Extract Mean Density}
\usage{
extract_mean_density(polygons, density_surface)
}
\arguments{
\item{polygons}{sf object with hotspot polygons}

\item{density_surface}{stars object with density surface}
}
\value{
numeric vector of mean density values
}
\description{
Extract Mean Density
}
