% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero_optimaLoc.R
\name{find_locations_centroid}
\alias{find_locations_centroid}
\title{Centroid based location finding (very fast)}
\usage{
find_locations_centroid(accidents, districts, num_locations)
}
\description{
Centroid based location finding (very fast)
}
\keyword{internal}
