% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspots.R
\name{plot.sero_hotspots}
\alias{plot.sero_hotspots}
\title{Plot method for sero_hotspots using ggplot2}
\usage{
\method{plot}{sero_hotspots}(x, ...)
}
\arguments{
\item{x}{sero_hotspots object}

\item{...}{additional arguments (unused)}
}
\value{
ggplot2 object
}
\description{
Plot method for sero_hotspots using ggplot2
}
