% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero_routes.R
\name{plot.sero_routes}
\alias{plot.sero_routes}
\title{Plot method for sero_routes using ggplot2}
\usage{
\method{plot}{sero_routes}(x, ...)
}
\arguments{
\item{x}{sero_routes object}

\item{...}{additional arguments (unused)}
}
\value{
ggplot2 object
}
\description{
Plot method for sero_routes using ggplot2
}
