% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspots.R
\name{print.sero_hotspots}
\alias{print.sero_hotspots}
\title{Print method for sero_hotspots}
\usage{
\method{print}{sero_hotspots}(x, ...)
}
\arguments{
\item{x}{sero_hotspots object}

\item{...}{additional arguments (unused)}
}
\description{
Print method for sero_hotspots
}
