% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero_optimal.R
\name{sero_export_results}
\alias{sero_export_results}
\title{Export optimal location results to various formats}
\usage{
sero_export_results(
  x,
  format = "csv",
  filename = "sero_optimal_locations",
  include_grid = FALSE
)
}
\arguments{
\item{x}{sero_optimal_locations object}

\item{format}{Export format ("csv", "geojson", "shapefile", "xlsx")}

\item{filename}{Output filename (without extension)}

\item{include_grid}{Logical, whether to include grid cells in export}
}
\value{
Character vector of created files
}
\description{
Export optimal location results to various formats
}
