% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero_visualization.R
\name{sero_plot_acc}
\alias{sero_plot_acc}
\title{Plot accidents with severity labels}
\usage{
sero_plot_acc(data, risk_categories = c(1, 2, 3), show_labels = TRUE)
}
\arguments{
\item{data}{List containing spatial data layers}

\item{risk_categories}{Vector of risk categories to include (default=c(1,2,3))}

\item{show_labels}{Logical, whether to show severity labels (default=TRUE)}
}
\value{
ggplot2 object
}
\description{
Plot accidents with severity labels
}
