% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero_visualization.R
\name{sero_plot_roads}
\alias{sero_plot_roads}
\title{Plot road network}
\usage{
sero_plot_roads(data, road_type = "all")
}
\arguments{
\item{data}{List containing spatial data layers}

\item{road_type}{Character, type of roads to highlight (default="all")}
}
\value{
ggplot2 object
}
\description{
Plot road network
}
