% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_mapping.R
\name{sero_save_optimal_locations}
\alias{sero_save_optimal_locations}
\title{Save optimal locations to spatialite database}
\usage{
sero_save_optimal_locations(optimal_locations, db_path)
}
\arguments{
\item{optimal_locations}{sf object with optimal locations}

\item{db_path}{path to sqlite database file}
}
\description{
Save optimal locations to spatialite database
}
