% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{sero_validate_data}
\alias{sero_validate_data}
\title{Validate and standardize spatial data for SERO analysis}
\usage{
sero_validate_data(data_list)
}
\arguments{
\item{data_list}{List of sf objects}
}
\value{
List of validated sf objects with standardized names
}
\description{
Validate and standardize spatial data for SERO analysis
}
